# docs/api-spec/openapi3.yaml
openapi: 3.0.1
info:
  title: Getchu API
  description: Getchu 프로젝트 API 명세서
  version: 1.0.0
servers:
- url: http://localhost:8080
...
paths:
...
  /members/{memberId}/reviews:
    get:
      tags:
      - Review
      summary: 리뷰 목록 조회
      description: |
        특정 회원이 받은 리뷰 목록을 커서 기반 페이지네이션으로 조회합니다.
        - 비로그인 상태에서도 조회 가능합니다.
        - 최신 리뷰가 먼저 반환됩니다.
        - 다음 페이지 조회 시 응답의 nextCursor 값을 cursor 파라미터로 전달합니다.
      operationId: review/get-reviews
      parameters:
      - name: memberId
        in: path
        description: 리뷰를 조회할 회원 ID
        required: true
        schema:
          type: string
      - name: cursor
        in: query
        description: 페이지 커서 (첫 페이지는 생략)
        required: false
        schema:
          type: string
      - name: size
        in: query
        description: "페이지 크기 (기본값: 10)"
        required: false
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewListResponse'
              examples:
                review/get-reviews:
                  value: "{\r\n  \"code\" : \"SUCCESS\",\r\n  \"message\" : \"요청이\
                    \ 성공적으로 처리되었습니다.\",\r\n  \"data\" : {\r\n    \"content\" : [ {\r\
                    \n      \"id\" : 2,\r\n      \"rating\" : 4.5,\r\n      \"content\"\
                    \ : \"친절하고 물건도 좋았습니다.\",\r\n      \"createdAt\" : \"2024-06-10T14:00:00\"\
                    ,\r\n      \"reviewerNickname\" : \"구매자닉네임\",\r\n      \"reviewerProfileImageUrl\"\
                    \ : \"https://cdn.example.com/buyer.jpg\",\r\n      \"productTitle\"\
                    \ : \"아이폰 15 팝니다\"\r\n    }, {\r\n      \"id\" : 1,\r\n      \"\
                    rating\" : 3.5,\r\n      \"content\" : \"무난한 거래였습니다.\",\r\n  \
                    \    \"createdAt\" : \"2024-05-20T09:30:00\",\r\n      \"reviewerNickname\"\
                    \ : \"다른구매자\",\r\n      \"reviewerProfileImageUrl\" : null,\r\n\
                    \      \"productTitle\" : \"맥북 에어 M2\"\r\n    } ],\r\n    \"nextCursor\"\
                    \ : \"eyJpZCI6MX0=\",\r\n    \"hasNext\" : true\r\n  }\r\n}"
  /trades/{tradeId}/reviews:
    post:
      tags:
      - Review
      summary: 리뷰 작성
      description: |
        거래 완료(SOLD) 상태인 거래에 대해 구매자가 판매자에게 리뷰를 작성합니다.
        - 거래당 리뷰는 1건만 작성 가능합니다.
        - 평점은 0.5 ~ 5.0 사이의 0.5 단위 값만 허용됩니다.
      operationId: review/create-review
      parameters:
      - name: tradeId
        in: path
        description: 리뷰를 작성할 거래 ID
        required: true
        schema:
          type: string
      - name: Authorization
        in: header
        description: Bearer JWT 액세스 토큰
        required: true
        schema:
          type: string
        example: Bearer test-token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewRequest'
            examples:
              review/create-review:
                value: "{\r\n  \"rating\" : 4.5,\r\n  \"content\" : \"친절하고 물건도 좋았습\
                  니다. 다음에 또 거래하고 싶어요.\"\r\n}"
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewCreateResponse'
              examples:
                review/create-review:
                  value: "{\r\n  \"code\" : \"SUCCESS\",\r\n  \"message\" : \"요청이\
                    \ 성공적으로 처리되었습니다.\"\r\n}"
components:
  schemas:
...
    ReviewRequest:
      title: ReviewRequest
      required:
      - content
      - rating
      type: object
      properties:
        rating:
          type: number
          description: "평점 (0.5 ~ 5.0, 0.5 단위)"
        content:
          type: string
          description: 리뷰 내용 (500자 이내)
...
    ReviewCreateResponse:
      title: ReviewCreateResponse
      required:
      - code
      - message
      type: object
      properties:
        code:
          type: string
          description: 응답 코드 (SUCCESS)
        message:
          type: string
          description: 응답 메시지
    ReviewListResponse:
      title: ReviewListResponse
      required:
      - code
      - message
      type: object
      properties:
        code:
          type: string
          description: 응답 코드 (SUCCESS)
        data:
          required:
          - hasNext
          type: object
          properties:
            nextCursor:
              type: string
              description: 다음 페이지 커서 (마지막 페이지면 null)
              nullable: true
            hasNext:
              type: boolean
              description: 다음 페이지 존재 여부
            content:
              type: array
              items:
                required:
                - content
                - createdAt
                - id
                - productTitle
                - rating
                - reviewerNickname
                type: object
                properties:
                  productTitle:
                    type: string
                    description: 거래된 상품명
                  createdAt:
                    type: string
                    description: 작성 일시
                  reviewerNickname:
                    type: string
                    description: 리뷰 작성자 닉네임
                  reviewerProfileImageUrl:
                    type: string
                    description: 리뷰 작성자 프로필 이미지 URL (없으면 null)
                    nullable: true
                  rating:
                    type: number
                    description: 평점 (0.5 ~ 5.0)
                  id:
                    type: number
                    description: 리뷰 ID
                  content:
                    type: string
                    description: 리뷰 내용
        message:
          type: string
          description: 응답 메시지